requirements.txt
pdfplumber
openai
pypdf2
reportlab
bs4
googlesearch-python
gspread
oauth2client